+++ extension/helper/cart.js
@@ -1,6 +1,8 @@
+const Message = require('../models/messages/message')
+
 /**
  * @param {int} sourceCartId
- * @param {object} context
+ * @param {SDKContext} context
  */
 const clearCart = (sourceCartId, context) => {
   const Shopify = require('../lib/shopify.api.js')(context.config, context.log)
@@ -18,7 +20,7 @@ const clearCart = (sourceCartId, context) => {
 /**
  * @param {object} updatedData
  * @param {int} targetCartId
- * @param {object} context
+ * @param {SDKContext} context
  */
 const updateCart = (updatedData, targetCartId, context) => {
   const Shopify = require('../lib/shopify.api.js')(context.config, context.log)
@@ -32,4 +34,53 @@ const updateCart = (updatedData, targetCartId, context) => {
   })
 }
 
-module.exports = { clearCart, updateCart }
+/**
+ * @param {Object} product
+ * @param {string} [product.customData]
+ * @returns {string|null}
+ * @throws {SyntaxError} If product.customData does not contain valid JSON.
+ */
+function extractVariantId (product) {
+  if (!product || !product.customData) return null
+
+  const customData = JSON.parse(product.customData)
+
+  return (customData && customData.variant_id) ? customData.variant_id : null
+}
+
+/**
+ * @param {Error} err
+ * @param {Object} err.errors
+ * @param {Object} err.errors.line_items
+ * @param {{code: string, message: string}[]} err.errors.line_items.[errorType]
+ *
+ * @return {{code: string, message: string, type: string}[]}
+ * @throws {Error} If err does not have an errors or error.line_items property
+ */
+function handleCartError (err) {
+  if (!err || !err.errors || !err.errors.line_items) throw err
+
+  const errorMessages = []
+  Object.values(err.errors.line_items).forEach(errorsPerLineItem => {
+    Object.entries(errorsPerLineItem).forEach(([errorType, errors]) => {
+      errors.forEach(error => {
+        let errorCode
+        switch (error.code) {
+          case 'not_enough_in_stock':
+            errorCode = 'EINSUFFICIENTSTOCK'
+            break
+          default:
+            errorCode = error.code
+        }
+
+        const errorMessage = new Message()
+        errorMessage.addErrorMessage(errorCode, error.message)
+        errorMessages.push(errorMessage.toJson())
+      })
+    })
+  })
+
+  return errorMessages
+}
+
+module.exports = { clearCart, updateCart, extractVariantId, handleCartError }